simplify and rename contract for testing -> msg_receiver
