+++ artifacts/checksums.txt
@@ -1,6 +1,7 @@
 ff08c9ca0765cc1ea3b884fbb71509c976445da43e9b73da1a2e0cb444e803f0  echo.wasm
 30e04bc90f355e818add3d8ad8b53a318f4411b44e6a7a6f45cb6435b8db7db4  hook_ibc_transfer.wasm
 c5fa166700dfe116e87b3ef3c66fa548feacff3bd6c9fed9e75a2f6072101686  ibc_transfer.wasm
+2d3ee79ffb877c8d6a5883a12800adaf3903675e0792421279c3c7170d2463f7  msg_receiver.wasm
 6d1a7342d73a3ae42fe7f5bc1473fc9af7361796d408467f77ddf0e657587f30  neutron_interchain_queries.wasm
 121579d06ca1d21131acba7c2ec07b31e91035b4e7c598c64548a1c71fc0bd28  neutron_interchain_txs.wasm
 0c2f7fb79d7986545536664576167271d85a3411e7c19ab553219312ddcca479  neutron_validators_test.wasm
