+++ artifacts/checksums_intermediate.txt
@@ -4,3 +4,4 @@ f73ffcb40e143b7ff155cb2ea211b2ec85bb3aeaf6009aacab0bb4491db21dc1  target/wasm32-
 004f49e56f1069218fa26a7d3192f6f4aa229649987e8eed07c3d0b91b73388e  target/wasm32-unknown-unknown/release/neutron_interchain_txs.wasm
 0cc3a704c7f4b1fb1bd8d0221e67757f6e5043cf0ac6fbe811ed00626096d44f  target/wasm32-unknown-unknown/release/neutron_validators_test.wasm
 eaca55e7dca6ce4eafbaec6895ab920e860f69b72c6a5b8afee8999c9e79cdfb  target/wasm32-unknown-unknown/release/reflect.wasm
+9a68687445b3e190eaa44f51892abe32ca2f3c6795746d11c2e1d21b26e78226  target/wasm32-unknown-unknown/release/msg_receiver.wasm
