--- contracts/hook_ibc_transfer/README.md
@@ -1,4 +1,0 @@
-# Neutron demo contracts
-## Hook IBC transfer contract to showcase ibc-hooks functionality (TODO: link to module docs)
-
-TODO: description
