--- contracts/hook_ibc_transfer/schema/execute_msg.json
@@ -1,116 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "ExecuteMsg",
-  "oneOf": [
-    {
-      "type": "object",
-      "required": [
-        "send"
-      ],
-      "properties": {
-        "send": {
-          "type": "object",
-          "required": [
-            "amount",
-            "channel",
-            "denom",
-            "to"
-          ],
-          "properties": {
-            "amount": {
-              "type": "integer",
-              "format": "uint128",
-              "minimum": 0.0
-            },
-            "channel": {
-              "type": "string"
-            },
-            "denom": {
-              "type": "string"
-            },
-            "memo": {
-              "type": [
-                "string",
-                "null"
-              ]
-            },
-            "timeout_height": {
-              "type": [
-                "integer",
-                "null"
-              ],
-              "format": "uint64",
-              "minimum": 0.0
-            },
-            "to": {
-              "type": "string"
-            }
-          }
-        }
-      },
-      "additionalProperties": false
-    },
-    {
-      "type": "object",
-      "required": [
-        "set_fees"
-      ],
-      "properties": {
-        "set_fees": {
-          "type": "object",
-          "required": [
-            "ack_fee",
-            "denom",
-            "recv_fee",
-            "timeout_fee"
-          ],
-          "properties": {
-            "ack_fee": {
-              "type": "integer",
-              "format": "uint128",
-              "minimum": 0.0
-            },
-            "denom": {
-              "type": "string"
-            },
-            "recv_fee": {
-              "type": "integer",
-              "format": "uint128",
-              "minimum": 0.0
-            },
-            "timeout_fee": {
-              "type": "integer",
-              "format": "uint128",
-              "minimum": 0.0
-            }
-          }
-        }
-      },
-      "additionalProperties": false
-    },
-    {
-      "type": "object",
-      "required": [
-        "test_msg"
-      ],
-      "properties": {
-        "test_msg": {
-          "type": "object",
-          "required": [
-            "arg",
-            "return_err"
-          ],
-          "properties": {
-            "arg": {
-              "type": "string"
-            },
-            "return_err": {
-              "type": "boolean"
-            }
-          }
-        }
-      },
-      "additionalProperties": false
-    }
-  ]
-}
