+++ contracts/msg_receiver/README.md
@@ -1,0 +1,5 @@
+# Description
+
+Simple contract message for integration tests.
+
+Can accept message test_arg and save info about incoming funds, sender and how many times was called.
