+++ contracts/msg_receiver/examples/schema.rs
@@ -17,7 +17,7 @@ use std::fs::create_dir_all;
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
-use hook_ibc_transfer::msg::{ExecuteMsg, InstantiateMsg};
+use msg_receiver::msg::{ExecuteMsg, InstantiateMsg};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
