+++ contracts/msg_receiver/schema/instantiate_msg.json
