+++ contracts/msg_receiver/src/lib.rs
